Move dataset class to datasets module
This is a start to move dataset lookup and management out of the django models module
