+++ budgetportal/bulk_upload.py
@@ -6,10 +6,12 @@ from budgetportal.models import (
     Sphere,
     Government,
     Department,
+)
+from datasets import (
     Dataset,
     Category,
     PackageDeletedException,
-    PackageWithoutGroupException
+    PackageWithoutGroupException,
 )
 from django import forms
 from django.contrib import messages
